data_ingestion/requirements.txt
snowflake-connector-python==4.4.0
python-dotenv==1.2.1
boto3==1.42.96
